создал базовый класс для task'ов; вынес в него общие методы и св-ва;
